+++ public/data/feedData.json
@@ -1,0 +1,27 @@
+{
+  "id": 1,
+  "userName": "피드 작성자",
+  "img": "이미지 URL",
+  "content": "피드 내용",
+  "isLiked": true,
+  "comment": [
+    {
+      "userName": "댓글 작성자1",
+      "content": "댓글 내용1",
+      "isLiked": true,
+      "isDeleted": true
+    },
+    {
+      "userName": "댓글 작성자2",
+      "content": "댓글 내용2",
+      "isLiked": true,
+      "isDeleted": true
+    },
+    {
+      "userName": "댓글 작성자3",
+      "content": "댓글 내용3",
+      "isLiked": true,
+      "isDeleted": true
+    }
+  ]
+}
