--- public/data/mockdata.md
@@ -1,1 +1,0 @@
-z
