--- src/pages/hyeongjunjeon/Main/Aside/AsideHyeongJun.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function Aside() {
-  return <div>Aside</div>;
-}
-
-export default Aside;
