+++ src/pages/hyeongjunjeon/Main/MainHyeongJun.scss
@@ -5,9 +5,11 @@ main {
   width: 100%;
   height: 100%;
   display: flex;
-  // justify-content: space-around;
-  // align-items: center;
   border: 1px solid white;
+
+  .ManyFeed {
+    width: 70%;
+  }
 
   article {
     width: 90%;
@@ -16,7 +18,7 @@ main {
   }
 
   .feeds {
-    width: 60%;
+    width: 100%;
     border: 1px solid white;
     margin: 5px auto 0px auto;
   }
